fix(content): clear session panel deep import gate
